+++ tests/weakrefs.py
@@ -1,6 +1,7 @@
 from django.utils import six
 
-from .app.models import Category, RegularCategory, Article, RegularArticle
+from .app.models import Category, RegularCategory, Article, RegularArticle,\
+    ArticleProxy
 
 from ._base import TestCase
 
@@ -9,14 +10,14 @@ class IdMapWeakRefsTests(TestCase):
 
     def setUp(self):
         n = 0
-        category = Category.objects.create(name="Category %d" % (n,))
-        regcategory = RegularCategory.objects.create(name="Category %d" % (n,))
+        category = Category.objects.create(name="Category %d" % n)
+        regcategory = RegularCategory.objects.create(name="Category %d" % n)
 
         for n in six.moves.xrange(0, 10):
-            Article.objects.create(name="Article %d" % (n,),
+            Article.objects.create(name="Article %d" % n,
                                    category=category,
                                    category2=regcategory)
-            RegularArticle.objects.create(name="Article %d" % (n,),
+            RegularArticle.objects.create(name="Article %d" % n,
                                           category=category,
                                           category2=regcategory)
 
@@ -60,3 +61,24 @@ class IdMapWeakRefsTests(TestCase):
         pk = article.pk
         article.delete()
         self.assertIsNone(Article.get_cached_instance(pk))
+
+    def test_proxy_model(self):
+        # check that access to instances of the same 'proxy' family is possible
+        # through all the classes of the 'proxy' family
+        article_list = list(Article.objects.all().select_related('category'))
+        proxy_list = []
+        for n in six.moves.xrange(0, 10):
+            a = ArticleProxy.objects.create(name="Proxy article %d" % n,
+                                            category=article_list[n].category,
+                                            category2=article_list[n].category2)
+            proxy_list.append(a)
+
+        for n in six.moves.xrange(0, 10):
+            article = ArticleProxy.get_cached_instance(article_list[n].pk)
+            self.assertTrue(isinstance(article, Article))
+            self.assertIs(article, article_list[n])
+
+        for n in six.moves.xrange(0, 10):
+            article = Article.get_cached_instance(proxy_list[n].pk)
+            self.assertTrue(isinstance(article, ArticleProxy))
+            self.assertIs(article, proxy_list[n])